Fix the compare for strings of different lengths
Previously, a shorter string was always assumed to be "less than" a
longer string, but this isn't necessarily true--e.g. "zzz" should be
"greater than" the string "aa" and this didn't match how strcmp()
normally works. This has been corrected such that the two strings are
compared such that the first n bytes are compared where n is the length
of the shorter string.